internal: drop hash checks from parsing step
It doesn't do anything useful, there are no other transactions on the network
tested but the ones we push to it. And we use len(blk.Transactions) for TPS
calculations already, so the only thing it changes is parsedCount, but that
just controls exit time and the value is the same, so nothing is changed.

This allows to remove hash map entirely, maintaining it quite expensive as
well as checking transactions against it while neo-bench should use the
minimum amount of CPU needed.